Refactor backend integration and prepare deployment

+++ src/main/java/com/upc/tp1inventory/DTO/DishIngredientDTO.java
@@ -2,15 +2,20 @@ package com.upc.tp1inventory.DTO;
 
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
 @Setter
 public class DishIngredientDTO {
+
     private UUID id;
+
     private UUID productId;
     private String productName;
-    private Double quantityNeeded;
+
+    private BigDecimal quantityNeeded;
     private String unit;
-    private Double costPerUnit;
+    private BigDecimal costPerUnit;
 }
